streamlit/requirements.txt
streamlit-cropper
Pillow